Correct check of returned status code for the API call
Looks like there was an error in checking the status code for API call. According to documentation request to `/me/submission_requests` will return status code 200 instead of 201. This bug makes the check incorrect and causes the script to try to create a new submission request instead of updating it.